+++ lib/cgi.rb
@@ -1,0 +1,12 @@
+class CGI
+  class << self
+    def escapeHTML(text)
+      text.
+        gsub('&', '&amp;').
+        gsub('<', '&lt;').
+        gsub('>', '&gt;').
+        gsub("'", '&#39;').
+        gsub('"', '&quot;')
+    end
+  end
+end